requirements.txt
psycopg2-binary==2.9.9
neo4j==5.15.0
python-dotenv==1.0.1 